+++ src/Adapter/Api/IliasRestApi.php
@@ -919,7 +919,7 @@ class IliasRestApi
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenById(int $id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenById(int $id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return $this->getObjectService()
             ->getChildrenById(
@@ -939,6 +939,7 @@ class IliasRestApi
                 $children_updated_after,
                 $children_updated_before,
                 $children_title,
+                $children_title_contains,
                 $ref_ids,
                 $in_trash
             );
@@ -949,7 +950,7 @@ class IliasRestApi
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByImportId(string $import_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByImportId(string $import_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return $this->getObjectService()
             ->getChildrenByImportId(
@@ -969,6 +970,7 @@ class IliasRestApi
                 $children_updated_after,
                 $children_updated_before,
                 $children_title,
+                $children_title_contains,
                 $ref_ids,
                 $in_trash
             );
@@ -979,7 +981,7 @@ class IliasRestApi
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByRefId(int $ref_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByRefId(int $ref_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return $this->getObjectService()
             ->getChildrenByRefId(
@@ -999,6 +1001,7 @@ class IliasRestApi
                 $children_updated_after,
                 $children_updated_before,
                 $children_title,
+                $children_title_contains,
                 $ref_ids,
                 $in_trash
             );
@@ -1401,7 +1404,7 @@ class IliasRestApi
      * @param ObjectType[]|null $types
      * @return ObjectDto[]
      */
-    public function getObjects(?int $id = null, ?string $import_id = null, int $ref_id = null, ?array $types = null, ?float $created = null, ?float $created_from = null, ?float $created_to = null, ?float $created_after = null, ?float $created_before = null, ?float $updated = null, ?float $updated_from = null, ?float $updated_to = null, ?float $updated_after = null, ?float $updated_before = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
+    public function getObjects(?int $id = null, ?string $import_id = null, int $ref_id = null, ?array $types = null, ?float $created = null, ?float $created_from = null, ?float $created_to = null, ?float $created_after = null, ?float $created_before = null, ?float $updated = null, ?float $updated_from = null, ?float $updated_to = null, ?float $updated_after = null, ?float $updated_before = null, ?string $title = null, ?string $title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : array
     {
         return $this->getObjectService()
             ->getObjects(
@@ -1420,6 +1423,7 @@ class IliasRestApi
                 $updated_after,
                 $updated_before,
                 $title,
+                $title_contains,
                 $ref_ids,
                 $in_trash
             );