+++ src/Adapter/Route/Object/GetChildren/GetChildrenByImportIdRoute.php
@@ -129,6 +129,11 @@ class GetChildrenByImportIdRoute implements Route
                     "Filter by title"
                 ),
                 RouteParamDocumentationDto::new(
+                    "title_contains",
+                    "string",
+                    "Filter title contains"
+                ),
+                RouteParamDocumentationDto::new(
                     "ref_ids",
                     "bool",
                     "Include ref ids"
@@ -217,6 +222,9 @@ class GetChildrenByImportIdRoute implements Route
             ),
             $request->getQueryParam(
                 "title"
+            ),
+            $request->getQueryParam(
+                "title_contains"
             ),
             $request->getQueryParam(
                 "ref_ids"