+++ src/Adapter/Route/Object/GetObjectsRoute.php
@@ -121,6 +121,11 @@ class GetObjectsRoute implements Route
                     "Filter by title"
                 ),
                 RouteParamDocumentationDto::new(
+                    "title_contains",
+                    "string",
+                    "Filter title contains"
+                ),
+                RouteParamDocumentationDto::new(
                     "ref_ids",
                     "bool",
                     "Include ref ids"
@@ -202,6 +207,9 @@ class GetObjectsRoute implements Route
                     ),
                     $request->getQueryParam(
                         "title"
+                    ),
+                    $request->getQueryParam(
+                        "title_contains"
                     ),
                     $request->getQueryParam(
                         "ref_ids"