+++ src/Service/Object/Command/GetChildrenCommand.php
@@ -38,7 +38,7 @@ class GetChildrenCommand
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenById(int $id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenById(int $id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         $object = $this->object_service->getObjectById(
             $id,
@@ -67,6 +67,7 @@ class GetChildrenCommand
             $children_updated_after,
             $children_updated_before,
             $children_title,
+            $children_title_contains,
             $ref_ids,
             $in_trash
         );
@@ -77,7 +78,7 @@ class GetChildrenCommand
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByImportId(string $import_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByImportId(string $import_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         $object = $this->object_service->getObjectByImportId(
             $import_id,
@@ -106,6 +107,7 @@ class GetChildrenCommand
             $children_updated_after,
             $children_updated_before,
             $children_title,
+            $children_title_contains,
             $ref_ids,
             $in_trash
         );
@@ -116,7 +118,7 @@ class GetChildrenCommand
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByRefId(int $ref_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByRefId(int $ref_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         $object = $this->object_service->getObjectByRefId(
             $ref_id,
@@ -145,6 +147,7 @@ class GetChildrenCommand
             $children_updated_after,
             $children_updated_before,
             $children_title,
+            $children_title_contains,
             $ref_ids,
             $in_trash
         );
@@ -155,7 +158,7 @@ class GetChildrenCommand
      * @param ObjectType[]|null $types
      * @return ObjectDto[]
      */
-    private function getChildren(?int $id = null, ?string $import_id = null, ?int $ref_id = null, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
+    private function getChildren(?int $id = null, ?string $import_id = null, ?int $ref_id = null, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?float $children_created = null, ?float $children_created_from = null, ?float $children_created_to = null, ?float $children_created_after = null, ?float $children_created_before = null, ?float $children_updated = null, ?float $children_updated_from = null, ?float $children_updated_to = null, ?float $children_updated_after = null, ?float $children_updated_before = null, ?string $children_title = null, ?string $children_title_contains = null, bool $ref_ids = false, ?bool $in_trash = null) : array
     {
         $objects = $this->ilias_database->fetchAll($this->ilias_database->query($this->getObjectChildrenQuery(
             $id,
@@ -176,6 +179,7 @@ class GetChildrenCommand
             $children_updated_after,
             $children_updated_before,
             $children_title,
+            $children_title_contains,
             null,
             $in_trash
         )));