+++ src/Service/Object/Command/GetObjectCommand.php
@@ -52,6 +52,7 @@ class GetObjectCommand
                 null,
                 null,
                 null,
+                null,
                 $in_trash
             )))) !== null) {
             if ($object !== null) {
@@ -74,6 +75,7 @@ class GetObjectCommand
         while (($object_ = $this->ilias_database->fetchAssoc($result ??= $this->ilias_database->query($this->getObjectQuery(
                 null,
                 $import_id,
+                null,
                 null,
                 null,
                 null,
@@ -111,6 +113,7 @@ class GetObjectCommand
                 null,
                 null,
                 $ref_id,
+                null,
                 null,
                 null,
                 null,